Fix signedness in calculation of memory throttling
I noticed that when I set the memrowset throttle threshold to larger than 4GB,
it started throttling immediately, rather than at the proper size.  This fixes
it to not overflow 32-bit ints.

Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5768
Tested-by: jenkins

+++ src/kudu/tablet/memrowset.cc
@@ -268,7 +268,8 @@ Status MemRowSet::CheckRowPresent(const RowSetKeyProbe &probe, bool *present,
 void MemRowSet::SlowMutators() {
   if (FLAGS_memrowset_throttle_mb == 0) return;
 
-  ssize_t over_mem = memory_footprint() - FLAGS_memrowset_throttle_mb * 1024 * 1024;
+  ssize_t over_mem = static_cast<int64_t>(memory_footprint()) -
+    static_cast<int64_t>(FLAGS_memrowset_throttle_mb) * 1024 * 1024;
   if (over_mem > 0) {
     if (!has_logged_throttling_ &&
         base::subtle::NoBarrier_AtomicExchange(&has_logged_throttling_, 1) == 0) {
